COG-628: UserEditDetailsService is extracted from Authorization service
