+++ src/main/java/gov/ca/cwds/idm/service/authorization/AuthorizationService.java
@@ -2,7 +2,6 @@ package gov.ca.cwds.idm.service.authorization;
 
 import com.amazonaws.services.cognitoidp.model.UserType;
 import gov.ca.cwds.idm.dto.User;
-import gov.ca.cwds.idm.dto.UserEditDetails;
 import gov.ca.cwds.idm.dto.UserUpdate;
 
 public interface AuthorizationService {
@@ -15,5 +14,9 @@ public interface AuthorizationService {
 
   void checkCanResendInvitationMessage(String id);
 
-  UserEditDetails getEditDetails(User user);
+  boolean canUpdateUser(String userId);
+
+  boolean canEditRoles(User user);
+
+  boolean canEditPermissions(User  user);
 }
